+++ include/logger.h
@@ -10,15 +10,25 @@ class PilesManager;
 
 class Logger {
   public:
-    Logger(bool logMoves);
+    enum LogType {
+      File,
+      Stdout,
+      None
+    };
+    Logger(LogType messagesLogType=None, LogType movesLogType=None, std::string messagesFileName="log.txt", std::string movesFileName="moves.txt");
     ~Logger();
     void error(std::string message, bool fatal=true);
     void logMove(int from, int to);
     void success(int moves);
+    void setMessagesLogType(LogType type, std::string messagesFileName="log.txt");
+    LogType getMessagesLogType() const;
+    void setMovesLogType(LogType type, std::string movesFileName="moves.txt");
+    LogType getMovesLogType() const;
   private:
-    std::ofstream m_logFile;
-    std::ofstream m_movesFile;
-    bool m_logMoves;
+    LogType m_messagesLogType;
+    std::ofstream m_messagesLogFile;
+    LogType m_movesLogType;
+    std::ofstream m_movesLogFile;
 };
 
 /*