+++ src/main.cpp
@@ -1,18 +1,17 @@
 #include <iostream>
 #include <string>
 #include <limits.h>  // for INT_MAX
-#include "pile.h"
-#include "disque.h"
-#include "piles_manager.h"
-#include "turn.h"
 #include "logger.h"
+#include "dmmaths.h"
 
 using namespace std;
 
 int main(int argc, char const *argv[]) {
   bool logMoves = false;
   if(argc >= 3) logMoves = true;
-  Logger logger(logMoves);
+  Logger logger(Logger::Stdout);
+  if(logMoves)
+    logger.setMovesLogType(Logger::File);
   if(argc < 2)
     logger.error("No argument", true);
   char *p;
@@ -24,48 +23,11 @@ int main(int argc, char const *argv[]) {
     logger.error("Argument isn't a good number", true);
   disques = conv;
 
-  Turn turns(disques);
-  PilesManager manager(disques, &turns, &logger);
+  DMMaths dmmaths;
+  dmmaths.setLogger(&logger);
+  dmmaths.setDisques(disques);
+  dmmaths.execute();
 
-  Disque *first = manager[0]->topDisque();
-  int turn;
-  Pile* dest;
-  int target, from;
-  bool last = false;
-  do {
-    turn = turns.turn();
-    if(turn+1 == disques) {
-      if(last) break;
-      last = true;
-      if(!manager.moveDisque(manager[0], manager[2])) logger.error("Couldn't move biggest disque", true);
-      if(turns.turn()+1 == disques) break;
-    }
-    else if(turn == 0) {
-      from = first->parent()->index();
-      target = 2-from;
-      if(manager[(from + 1) % PILES]->empty() &&
-        manager[(from + 2) % PILES]->empty());
-      else if(!manager[(from + 1) % PILES]->empty() &&
-        (manager[(from + 2) % PILES]->empty() ||
-          manager[(from + 1) % PILES]->topDisque()->size() <
-          manager[(from + 2) % PILES]->topDisque()->size()))
-            target = 1;
-      else target = 2;
-
-      if(first->parent()->even()) {
-        target = target % 2 + 1;
-      }
-      dest = manager[(from + target) % PILES];
-      if(!manager.moveDisque(first->parent(), dest))
-        logger.error("Couldn't move first disque", true);
-    }
-    else {
-      if(!manager.moveDisque(manager.disque(turn)->parent(), manager[(manager.disque(turn)->parent()->index()+1)%PILES]))
-        if(!manager.moveDisque(manager.disque(turn)->parent(), manager[(manager.disque(turn)->parent()->index()+2)%PILES]))
-          logger.error("Couldn't move disque anywhere", true);
-    }
-  } while(1);
-
-  logger.success(manager.moves());
+  logger.success(dmmaths.getCounter());
   return EXIT_SUCCESS;
 }